pyproject.toml
[project]
name = "adm-aos"
description = "Administer's AOS service for distributing apps."
readme = "README.md"

version = "5.6"
requires-python = ">=3.12"

authors = [{ name = "pyxfluff", email = "pyx@admsoftware.org" }]
dependencies = [
    # HTTP
    "uvicorn",
    "fastapi",
    "httpx",
    # Utilities
    "Levenshtein",
    "colorthief",
    "pymongo",
    "pathlib",
    "nanoid",
    "orjson",
    "psutil>=7.0.0",
    "rich",
    "pydantic",
    # Discord
    #"audioop-lts",
    #"py-cord",
    # Development
    "matplotlib",
    "flake8>=7.2.0",
    "libsass>=0.23.0",
    "jinja2>=3.1.6",
    "argon2-cffi>=25.1.0",
    "bcrypt>=4.3.0",
    "ruff>=0.12.7",
]

urls.homepage = "https://admsoftware.org"
urls.repository = "https://github.com/administer-org/app-server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch]
metadata.allow-direct-references = true
build.targets.wheel.packages = ["AOS"]

[project.scripts]
aos = "AOS.__main__:main"

[tool.ruff.format]
quote-style = "double"
line-ending = "lf"
indent-style = "space"
skip-magic-trailing-comma = true
